+++ src/Main.java
@@ -1,4 +1,7 @@
 import enums.StatusTask;
+import managers.HistoryManager;
+import managers.Managers;
+import managers.TaskManager;
 import tasks.Epic;
 import tasks.Subtask;
 import tasks.Task;
