+++ src/handlers/HistoryHandler.java
@@ -1,6 +1,9 @@
+package handlers;
+
 import com.google.gson.*;
 import com.sun.net.httpserver.HttpExchange;
 import enums.Endpoint;
+import managers.TaskManager;
 import tasks.Task;
 
 import java.io.*;
