+++ src/handlers/TasksHandler.java
@@ -1,50 +1,53 @@
+package handlers;
+
 import com.google.gson.*;
 import com.sun.net.httpserver.HttpExchange;
 import enums.Endpoint;
-import tasks.Subtask;
+import managers.Managers;
+import managers.TaskManager;
+import tasks.Task;
 
 import java.io.*;
 import java.util.*;
 
-public class SubtasksHandler extends BaseHttpHandler {
+public class TasksHandler extends BaseHttpHandler {
 
-    public SubtasksHandler(TaskManager taskManager) {
+    public TasksHandler(TaskManager taskManager) {
         super(taskManager);
     }
 
-    private Subtask updateSubtaskPost(JsonObject jsonObject) {
+    private Task updateTaskPost(JsonObject jsonObject) {
         int id = jsonObject.get("id").getAsInt();
-        Subtask subtask = taskManager.getSubtaskByID(id);
+        Task task = taskManager.getTaskByID(id);
         if (jsonObject.has("title")) {
-            subtask.setTitle(jsonObject.get("title").getAsString());
+            task.setTitle(jsonObject.get("title").getAsString());
         }
         if (jsonObject.has("description")) {
-            subtask.setDescription(jsonObject.get("description").getAsString());
+            task.setDescription(jsonObject.get("description").getAsString());
         }
-        taskManager.updateSubtask(subtask);
-        return subtask;
+        taskManager.updateTask(task);
+        return task;
     }
 
-    private Subtask createSubtaskPost(JsonObject jsonObject) {
+    private Task createTaskPost(JsonObject jsonObject) {
         String inputTitle = jsonObject.get("title").getAsString();
         String inputDescription = jsonObject.get("description").getAsString();
-        int inputEpicID = jsonObject.get("epicId").getAsInt();
         int inputMinutes = jsonObject.get("minutes").getAsInt();
         String inputDate = jsonObject.get("date").getAsString();
-        Subtask subtask = new Subtask(inputTitle, inputDescription, inputEpicID, inputMinutes, Managers.stringToDate(inputDate));
-        if (taskManager.isValid(subtask)) {
-            taskManager.addSubtask(subtask);
+        Task task = new Task(inputTitle, inputDescription, inputMinutes, Managers.stringToDate(inputDate));
+        if (taskManager.isValid(task)) {
+            taskManager.addTask(task);
         } else {
-            subtask = null;
+            task = null;
         }
-        return subtask;
+        return task;
     }
 
     @Override
     public void handle(HttpExchange exchange) throws IOException {
         int id;
-        Subtask subtask;
-        List<Subtask> subtasks;
+        Task task;
+        List<Task> tasks;
         Optional<Integer> optionVal;
         try {
             String method = exchange.getRequestMethod();
@@ -52,21 +55,21 @@ public class SubtasksHandler extends BaseHttpHandler {
             Endpoint endpoint = getEndpoint(path, method);
             Gson gson = getGson();
             switch (endpoint) {
-                case GET_SUBTASKS:
-                    subtasks = taskManager.getSubtasks();
-                    writeResponse(exchange, gson.toJson(subtasks), 200);
+                case GET_TASKS:
+                    tasks = taskManager.getTasks();
+                    writeResponse(exchange, gson.toJson(tasks), 200);
                     break;
-                case GET_SUBTASK:
+                case GET_TASK:
                     optionVal = getPathParam(exchange);
                     if (optionVal.isEmpty()) {
-                        writeResponse(exchange, "Некорректный идентификатор подзадачи", 400);
+                        writeResponse(exchange, "Некорректный идентификатор задачи", 400);
                         break;
                     }
                     id = optionVal.get();
-                    subtask = taskManager.getSubtaskByID(id);
-                    writeResponse(exchange, gson.toJson(subtask), 200);
+                    task = taskManager.getTaskByID(id);
+                    writeResponse(exchange, gson.toJson(task), 200);
                     break;
-                case POST_SUBTASKS:
+                case POST_TASKS:
                     JsonElement jsonElement = getRequestBody(exchange);
                     if (jsonElement == null || !jsonElement.isJsonObject()) {
                         writeResponse(exchange, "Не удалось извлечь тело запроса", 400);
@@ -74,8 +77,8 @@ public class SubtasksHandler extends BaseHttpHandler {
                     }
                     JsonObject jsonObject = jsonElement.getAsJsonObject();
                     if (jsonObject.has("id")) {
-                        subtask = updateSubtaskPost(jsonObject);
-                        writeResponse(exchange, gson.toJson(subtask), 200);
+                        task = updateTaskPost(jsonObject);
+                        writeResponse(exchange, gson.toJson(task), 200);
                     } else {
                         if (!jsonObject.has("title")) {
                             writeResponse(exchange, "Из тела запроса не удалось извлечь обязательный параметр title", 404);
@@ -93,24 +96,24 @@ public class SubtasksHandler extends BaseHttpHandler {
                             writeResponse(exchange, "Из тела запроса не удалось извлечь обязательный параметр date", 404);
                             break;
                         }
-                        subtask = createSubtaskPost(jsonObject);
-                        if (subtask != null) {
-                            writeResponse(exchange, gson.toJson(subtask), 200);
+                        task = createTaskPost(jsonObject);
+                        if (task != null) {
+                            writeResponse(exchange, gson.toJson(task), 200);
                         } else {
-                            writeResponse(exchange, "Добавляемая подзадача пересекается с существующей", 406);
+                            writeResponse(exchange, "Добавляемая задача пересекается с существующей", 406);
                         }
                     }
                     break;
-                case DELETE_SUBTASK:
+                case DELETE_TASK:
                     optionVal = getPathParam(exchange);
                     if (optionVal.isEmpty()) {
                         writeResponse(exchange, "Некорректный идентификатор задачи", 400);
-                    } else {
-                        id = optionVal.get();
-                        taskManager.getSubtaskByID(id);
-                        taskManager.removeSubtaskByID(id);
-                        writeResponse(exchange, null, 201);
+                        break;
                     }
+                    id = optionVal.get();
+                    taskManager.getTaskByID(id);
+                    taskManager.removeTaskByID(id);
+                    writeResponse(exchange, null, 201);
                     break;
                 default:
                     writeResponse(exchange, "Сервер не обслуживает эндпоинт " + method + path, 501);
