+++ src/handlers/TypeAdapterDuration.java
@@ -1,3 +1,5 @@
+package handlers;
+
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonToken;
