+++ src/managers/HistoryManager.java
@@ -1,3 +1,5 @@
+package managers;
+
 import tasks.Task;
 
 import java.util.List;
