+++ src/managers/InMemoryTaskManager.java
@@ -1,3 +1,5 @@
+package managers;
+
 import enums.StatusTask;
 import tasks.Epic;
 import tasks.Subtask;
