+++ src/managers/Node.java
@@ -1,3 +1,5 @@
+package managers;
+
 public class Node<T> {
 
     private final T data;
