+++ test/USprint8Task1Test.java
@@ -1,4 +1,7 @@
 import enums.StatusTask;
+import managers.HistoryManager;
+import managers.Managers;
+import managers.TaskManager;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.BeforeAll;
